Brownian.jl

Documentation for Brownian.jl

Brownian.rand_chol — Function

rand_chol

generates FBM using the method based on Cholesky decomposition. T. Dieker, Simulation of Fractional Brownian Motion, master thesis, 2004. The complexity of the algorithm is O(n^3), where n is the number of FBM samples.

Brownian.rand_fft — Function

rand_fft

generates FBM using fast Fourier transform (FFT). The time interval of FBM is [0, 1] with a stepsize of 2^p, where p is a natural number. The algorithm is known as the Davies-Harte method or the method of circular embedding. R.B. Davies and D.S. Harte, Tests for Hurst Effect, Biometrika, 74 (1), 1987, pp. 95-102. For a more recent publication, see P.F. Craigmile, Simulating a Class of Stationary Gaussian Processes Using the Davies–Harte Algorithm, With Application to Long Memory Processes, Journal of Time Series Analysis, 24 (5), 2003, pp. 505-511. The complexity of the algorithm is O(n*log(n)), where n=2^p is the number of FBM samples.

Brownian.rand_rl — Function

rand_rl

Implementation of "exact discrete" method for simulating Riemann-Liouville fBm Based on Muniandy, S. & Lim, S. Modeling of locally self-similar processes using multifractional Brownian motion of Riemann-Liouville type. Physical Review E 63, 046104. ISSN: 1063-651X (2001). Specifically, Eqn. 17, 18 and 19. From Muniandy (2001), Reimann-Liouville fBm is defined as,

\[B_H(t)=\frac{1}{\Gamma(H+0.5)}\int^{t}_{0}(t-s)^(H-0.5)dB(s), t\geq0.\]

Here we focus on the discrete time $t_j=j\Delta t$ approximation,

\[B_H(t_j)=\frac{1}{\Gamma(H+0.5)}\sum^{j}_{i=1}\int^{i\Delta t}_{(i-1)\Delta t}(t_j-\tau)^(H-0.5)dB(\tau)\]

The exact solution to the interior integral results in a weighting function, implemented here with the :Exact key. An "improved" weighting function was proposed by Rambaldi, S. & Pinazza, O. An accurate fractional Brownian motion generator. Physica A 208, 21–30 (1994). We have implemented this here with the :Improved key.